% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ecoligenomics.R
\docType{package}
\name{Ecoligenomics}
\alias{Ecoligenomics}
\title{Ecoligenomics package}
\description{
This package can be used in order to type (using E coli MLST scheme) and screen (e.g. for the presence of virulence) E coli genomes. The genomes can be either complete or draft (i.e., multiple contigs)
}
\author{
Jerome Ambroise \email{jerome.ambroise@uclouvain.be}
}
