% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type-MLST.R
\name{typemlst}
\alias{typemlst}
\title{type MLST of a draft genome}
\usage{
typemlst(genomePath, type)
}
\arguments{
\item{reference}{the reference sequence that you want to screen. Fasta file in one or severa sequences}

\item{querry}{the querry sequence. Fasta file in one or severa sequences}

\item{dir.out}{the directory of the output}
}
\value{
numeric value of the percentage of the reference sequence which is covered by the querry
}
\description{
You provide the path of the draft genome and
}
